requirements.txt
fire==0.5.0
lmdb==1.2.1
matplotlib==3.7.1
natsort==7.1.1
nltk==3.6.2
numpy==1.20.0
opencv_contrib_python==4.5.1.48
opencv_python==4.5.1.48
Pillow==9.5.0
Pillow==10.1.0
scikit_image==0.18.0
scipy==1.9.1
setuptools==68.0.0
six==1.16.0
strsimpy==0.2.1
tabulate==0.9.0
thop==0.1.1.post2209072238
torch==1.10.1+cu111
torchvision==0.11.2+cu111
transformers==4.2.1
validators==0.20.0
Wand==0.6.7
